Raumautomation 2Arduino Raspberry/pi_backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pyserial==3.5
simple-pid==2.0.0
flask-cors==4.0.0
